bibtex-to-rdf-v2/requirement.txt
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
latexcodec==2.0.1
MarkupSafe==2.1.1
pybtex==0.24.0
pyparsing==3.0.9
python-dotenv==0.20.0
PyYAML==6.0
rdflib==6.1.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
